refactor: extract quote service boundary
