+++ crates/runtime/src/services/mod.rs
@@ -1,6 +1,10 @@
+pub(crate) mod blocking_simulation_executor;
 pub mod broadcaster;
 pub mod broadcaster_sessions;
 pub mod broadcaster_subscription;
 pub mod encode;
+pub mod quote_service;
 pub mod quotes;
 pub mod stream_builder;
+
+pub use quote_service::QuoteService;
